chore: extract hook config & schema to own file
