+++ internal/config/recipe_manifest.go
@@ -37,20 +37,10 @@ var (
 			"hooks": map[string]any{
 				"type": "object",
 				"properties": map[string]any{
-					"before":     map[string]any{"$ref": "#/$defs/hook"},
-					"after":      map[string]any{"$ref": "#/$defs/hook"},
-					"on-success": map[string]any{"$ref": "#/$defs/hook"},
-					"on-failure": map[string]any{"$ref": "#/$defs/hook"},
-				},
-			},
-		},
-		"$defs": map[string]any{
-			"hook": map[string]any{
-				"type":     "object",
-				"required": []any{"shell", "command"},
-				"properties": map[string]any{
-					"shell":   map[string]any{"enum": []any{"bash", "sh"}},
-					"command": map[string]any{"type": "string"},
+					"before":     hookSchemaRef,
+					"after":      hookSchemaRef,
+					"on-success": hookSchemaRef,
+					"on-failure": hookSchemaRef,
 				},
 			},
 		},
@@ -61,6 +51,10 @@ var (
 func loadRecipeManifestV1Schema() (*jsonschema.Schema, error) {
 	compiler := jsonschema.NewCompiler()
 	err := compiler.AddResource(recipeManifestV1SchemaUrl, _recipeManifestV1Schema)
+	if err != nil {
+		return nil, err
+	}
+	err = addHookSchema(compiler)
 	if err != nil {
 		return nil, err
 	}
@@ -77,11 +71,6 @@ func init() {
 		log.Panicf("[internal error] failed to load recipe manifest v1 schema: %v", err)
 	}
 	recipeManifestV1Schema = *res
-}
-
-type HookV1 struct {
-	Shell   string `mapstructure:"shell"`
-	Command string `mapstructure:"command"`
 }
 
 type HooksV1 struct {
